requirements.txt
streamlit
PyPDF2
pdf2image
pytesseract
spacy
scikit-learn
pandas
numpy
beautifulsoup4
requests
python-dotenv
openai
bertopic
umap-learn
google-generativeai
hdbscan
googlesearch-python
Pillow
